Add files via upload
The purpose of the exercise is to identify the extent of credit risk migration (change in the risk profile of the portfolio) between the two periods (Period 1 and Period 2). The candidate has a number of alternatives for defining the migration impact and should provide not only a result in terms of risk-weight (RWA/EAD), probability of default, and expected loss, but also consider the impact of new and existing customers and describe the impact of defaulted and unrated customers and whether they are relevant to the analysis by considering the effects on PD, RWA, and EL. Which customers have increased/decreased capital consumption the most? Is the rating and PD value consistent?